ability to add and remove style declarations
